+++ packages/paste-core/primitives/tabs/src/index.tsx
@@ -1,0 +1,6 @@
+export {
+  useTabState as useTabsPrimitiveState,
+  Tab as TabsPrimitive,
+  TabList as TabsPrimitiveList,
+  TabPanel as TabsPrimitivePanel,
+} from '@twilio-paste/reakit-library';