+++ packages/paste-core/components/heading/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.8](https://github.com/twilio-labs/paste/compare/@twilio-paste/heading@2.0.7...@twilio-paste/heading@2.0.8) (2020-06-01)
+
+**Note:** Version bump only for package @twilio-paste/heading
+
+
+
+
+
 ## [2.0.7](https://github.com/twilio-labs/paste/compare/@twilio-paste/heading@2.0.6...@twilio-paste/heading@2.0.7) (2020-06-01)
 
 **Note:** Version bump only for package @twilio-paste/heading